수정 + 기능 3 구현
