--- myMovieReviews/myreviews/reviews/admin.py
@@ -1,4 +1,0 @@
-from django.contrib import admin
-from .models import Review
-# Register your models here.
-admin.site.register(Review)
