--- myMovieReviews/myreviews/reviews/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path('', views.review_list, name='review_list'),
-    path('<int:pk>/', views.review_detail, name='review_detail'),
-]
