+++ myMovieReviews/reviews/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from .models import Review
+# Register your models here.
+
+admin.site.register(Review)
