+++ myMovieReviews/reviews/templates/reviews/review_form.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <title>리뷰 작성</title>
+    </head>
+    <body>
+        <h1>리뷰 작성</h1>
+        <form method="post" enctype="multipart/form-data">
+            {% csrf_token %}
+            {{ form.as_p }}
+            <button type="submit">저장</button>
+        </form>
+    </body>
+</html>
