+++ myMovieReviews/reviews/views.py
@@ -1,0 +1,21 @@
+from django.shortcuts import render,  get_object_or_404, redirect
+from .models import Review
+from .forms import ReviewForm
+# Create your views here.
+def review_list(request):
+    reviews = Review.objects.all()
+    return render(request, 'reviews/review_list.html', {'reviews': reviews})
+
+def review_detail(request, pk): 
+    review = get_object_or_404(Review, pk=pk)
+    return render(request, 'reviews/review_detail.html', {'review': review})
+
+def review_create(request):
+    if request.method == 'POST':
+        form = ReviewForm(request.POST, request.FILES)
+        if form.is_valid():
+            form.save()
+            return redirect('review_list')
+    else:
+        form = ReviewForm()
+    return render(request, 'reviews/review_form.html', {'form': form})
